fix issue: DecimalRow formatter wrong for too-big number?

+++ Source/Rows.swift
@@ -355,7 +355,7 @@ public class _PasswordRow: FieldRow<String, PasswordCell> {
     }
 }
 
-public class _DecimalRow: FieldRow<Float, DecimalCell> {
+public class _DecimalRow: FieldRow<Double, DecimalCell> {
     public required init(tag: String?) {
         super.init(tag: tag)
         let numberFormatter = NSNumberFormatter()
